+++ armsrc/hfsnoop.c
@@ -14,8 +14,7 @@
 #include "BigBuf.h"
 #include "util.h"
 #include "apps.h"
-#include "cmd.h"
-#include "usb_cdc.h"	// for usb_poll_validate_length
+#include "usb_cdc.h"
 #include "fpga.h"
 #include "fpgaloader.h"
