+++ armsrc/mifaresniff.c
@@ -18,7 +18,7 @@
 #include "crapto1/crapto1.h"
 #include "mifareutil.h"
 #include "common.h"
-#include "cmd.h"
+#include "usb_cdc.h"
 #include "BigBuf.h"
 #include "fpgaloader.h"
